Add files via upload
In this project I wrote two simple scripts that transforms on single img or video into a new img or video designe to be displayed with a transparent tetrahedron in order to represent a simulation of a 3d hologram (I will post and link a video so you can see what I'm talking about). In this folder you will find one img and one video to test both scripts.