Add files via upload
A movement detector created in python mainly using open-cv library, connected to a telegram bot that sends you a 2 sec gif when movement is detected. You can also control a few basic features of the script using this bot with some commands.
The pdf file you will find represents a flow chart diagram that represents the functionality of the whole code.